requirements.txt
djangorestframework
openai
pyairtable